commit_3

+++ src/main/resources/db.changelog/tablesDB.sql
@@ -3,10 +3,10 @@
 
 CREATE TABLE IF NOT EXISTS developers_database.developers
 (
-    id           LONG         NOT NULL AUTO_INCREMENT,
+    id           INTEGER      NOT NULL AUTO_INCREMENT,
     first_name   VARCHAR(100) NOT NULL,
     last_name    VARCHAR(100) NOT NULL,
-    specialty_id LONG         NOT NULL,
+    specialty_id INTEGER      NOT NULL,
     status_name  VARCHAR(100) NOT NULL,
     PRIMARY KEY (id)
 );
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS developers_database.developers
 
 CREATE TABLE IF NOT EXISTS developers_database.specialties
 (
-    id             LONG         NOT NULL AUTO_INCREMENT,
+    id             INTEGER      NOT NULL AUTO_INCREMENT,
     specialty_name VARCHAR(100) NOT NULL,
     status_name    VARCHAR(100) NOT NULL,
     PRIMARY KEY (id)
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS developers_database.specialties
 
 CREATE TABLE IF NOT EXISTS developers_database.skills
 (
-    id          LONG         NOT NULL AUTO_INCREMENT,
+    id          INTEGER      NOT NULL AUTO_INCREMENT,
     skill_name  VARCHAR(100) NOT NULL,
     status_name VARCHAR(100) NOT NULL,
     PRIMARY KEY (id)
@@ -32,8 +32,8 @@ CREATE TABLE IF NOT EXISTS developers_database.skills
 
 CREATE TABLE IF NOT EXISTS developers_database.developer_skills
 (
-    developer_id LONG DEFAULT NULL,
-    skill_id     LONG DEFAULT NULL,
+    developer_id INTEGER DEFAULT NULL,
+    skill_id     INTEGER DEFAULT NULL,
     CONSTRAINT `FK_developer_skills_developers` FOREIGN KEY (developer_id) REFERENCES developers (id) ON DELETE CASCADE,
     CONSTRAINT `FK_developer_skills_skills` FOREIGN KEY (skill_id) REFERENCES skills (id) ON DELETE CASCADE
 );
